+++ ninjaphenix-container-library/src/main/java/ninjaphenix/containerlib/impl/client/screen/SingleScreen.java
@@ -24,7 +24,7 @@ public class SingleScreen<T extends SingleScreenHandler> extends AbstractScreen<
     {
         super.init();
         int settingsXOffset = -19;
-        if(FabricLoader.getInstance().isModLoaded("inventorysorter")) { settingsXOffset -= 18; }
+        if (FabricLoader.getInstance().isModLoaded("inventorysorter")) { settingsXOffset -= 18; }
         addButton(new ScreenTypeSelectionScreenButton(x + backgroundWidth + settingsXOffset, y + 4,
                 (button, matrices, mouseX, mouseY) -> renderTooltip(matrices, button.getMessage(), mouseX, mouseY)));
         final int blanked = SCREEN_META.BLANK_SLOTS;